Added Makefile also added a testing source code it's planning the setup process of streams (media), putting it all together it launches a (de)queueing thread for v4l2 brings up the isp configures speaker and mic then records on a file and plays another in loop ctrl+c is shutting down nice, the kernel is freeing right an interesting and painful thingy is that first run a kernel message appears and only 10fps are calculated from camera after first ctrl+c and re-run, it arrives to 15fps
this code still not dumping encoding video frames
we can take the raw frames but are not usable:
1) couldn't see them fine (green or crashed images); maybe choosing wrong format on ffplay, or maybe wrong isp setup (haven't reviewed with patience
the results of the AI; still ordering things that it can't alone, and trying to setting up in logical order (open_devs(), etc-- vendor ipc is a crazy
rocket launcher of system calls with altered order which seems not always to be ok), i mean, if you don't open /dev/video0 before setting up the isp,
don't try; they "work together", even 1 is isp and other (char device) is video0 (v4l2...).
2) video encoder (H264 hardware encoding) is the logical answer: the device is very weak to work with 2.5MB/s data; not possible to save on sd,
not possible to send over wireless (at least in my work scenario), not possible to software encoding, so, that's why probably we will have to work with
encoded frames AND the reason why it uses DMA buffers shared over all the streaming hardware-- not the speaker i guess...

the testing code sent here is to get an idea of how to start thinking with streaming but still didn't implemented the uio0 part nor the well traceable
cleaning and shutdown.

anyway, the original ipc reboots the system if killed so... maybe there's no a clean solution to shutdown the daemon and later bringing it up again
hoping it to work nice, without modifying anything in kernel or worstly not possible due to hardware design limitations-- i'm not sure.